+++ scripts/build-storybook-prerelease.sh
@@ -5,7 +5,7 @@ cd "${TRAVIS_BUILD_DIR}"
 mkdir -p ../tmp
 
 cd "${TRAVIS_BUILD_DIR}/packages/docs"
-yarn build -o "${TRAVIS_BUILD_DIR}/../tmp/v${TRAVIS_BRANCH}" --ci --quiet
+yarn build -o "${TRAVIS_BUILD_DIR}/../tmp/v${TRAVIS_BRANCH}" --quiet
 
 cd "${TRAVIS_BUILD_DIR}"
 git checkout "${TRAVIS_BRANCH}-docs"