+++ scripts/build-storybook.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 set -e
 
-yarn build:libs
+echo "STORYBOOK Opening project root directory..."
+cd "$TRAVIS_BUILD_DIR"
 
+echo "STORYBOOK Fetching packages version..."
 VERSION=`cat lerna.json \
   | grep version \
   | head -1 \
@@ -11,29 +13,44 @@ VERSION=`cat lerna.json \
   | tr -d '[[:space:]]'n`
 
 if [ -z "${VERSION}" ]; then
-  echo "Null release version! Can not build Storybook."
+  echo "STORYBOOK Can not build Storybook. Could not find version information at lerna.json file."
   exit 0
+else
+  echo "STORYBOOK Generating Storybook for version ${VERSION}..."
 fi
 
-yarn lerna run test:ci
-
-cd "$TRAVIS_BUILD_DIR"
+echo "STORYBOOK Creating temp directory for bundle..."
 mkdir -p ../tmp
 
+echo "STORYBOOK Opening docs package directory..."
 cd "$TRAVIS_BUILD_DIR/packages/docs"
-yarn build -o "${TRAVIS_BUILD_DIR}/../tmp/v${VERSION}" --ci --quiet
 
+echo "STORYBOOK Generating JSON report for Jest addon..."
+yarn lerna run test:packages
+
+echo "STORYBOOK Building Storybook..."
+yarn build -o "${TRAVIS_BUILD_DIR}/../tmp/v${VERSION}" --quiet
+
+echo "STORYBOOK Going back to project root directory..."
 cd "$TRAVIS_BUILD_DIR"
+
+echo "STORYBOOK Moving to docs branch..."
 git checkout .
 git checkout master-docs
 
+echo "STORYBOOK Adding version ${VERSION} to versions JSON file..."
 cd scripts
-node helpers/addVersionOnConfig.js "v${VERSION}"
+node helpers/addVersionOnConfig.js "${VERSION}"
 
+echo "STORYBOOK Going back to project root directory..."
 cd "${TRAVIS_BUILD_DIR}"
+
+echo "STORYBOOK Copying generated Storybook to releases directory..."
 cp -r "${TRAVIS_BUILD_DIR}/../tmp/v${VERSION}" packages/docs/dist/releases
 
+echo "STORYBOOK Committing changes..."
 git add --all
 git commit -m "docs: generating storybook for version ${VERSION} [skip ci]"
 
+echo "STORYBOOK Updating master-docs branch..."
 git push -f -u origin master-docs