+++ packages/web/CHANGELOG.md
@@ -3,6 +3,127 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.9.0](https://github.com/natura-cosmeticos/natds-js/compare/v0.9.0-alpha.DSY-1219.0...v0.9.0) (2020-07-09)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
+## [0.8.11](https://github.com/natura-cosmeticos/natds-js/compare/v0.8.10...v0.8.11) (2020-07-08)
+
+
+### Bug Fixes
+
+* **deps:** update dependency lodash to v4.17.19 ([55ef4e6](https://github.com/natura-cosmeticos/natds-js/commit/55ef4e63945cff0a7c6d096c4e75f861359b3948))
+
+
+
+
+
+## [0.8.10](https://github.com/natura-cosmeticos/natds-js/compare/v0.8.9...v0.8.10) (2020-07-08)
+
+
+### Bug Fixes
+
+* **deps:** update dependency lodash to v4.17.18 ([4c12ed8](https://github.com/natura-cosmeticos/natds-js/commit/4c12ed8b7a6bf11b17ebcfe51bfbb2db78a205d1))
+
+
+
+
+
+## [0.8.9](https://github.com/natura-cosmeticos/natds-js/compare/v0.8.8...v0.8.9) (2020-07-08)
+
+
+### Bug Fixes
+
+* **deps:** update dependency lodash to v4.17.17 ([5f74fe6](https://github.com/natura-cosmeticos/natds-js/commit/5f74fe6fba400ee7dfc968955e0b2a2d4d6c10c9))
+
+
+
+
+
+## [0.8.8](https://github.com/natura-cosmeticos/natds-js/compare/v0.8.7...v0.8.8) (2020-07-07)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
+## [0.8.6](https://github.com/natura-cosmeticos/natds-js/compare/v0.8.5...v0.8.6) (2020-07-07)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
+## [0.8.5](https://github.com/natura-cosmeticos/natds-js/compare/v0.8.4...v0.8.5) (2020-07-07)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
+## [0.8.4](https://github.com/natura-cosmeticos/natds-js/compare/v0.8.3...v0.8.4) (2020-07-06)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
+## [0.8.3](https://github.com/natura-cosmeticos/natds-js/compare/v0.8.2...v0.8.3) (2020-07-06)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
+## [0.8.2](https://github.com/natura-cosmeticos/natds-js/compare/v0.8.1...v0.8.2) (2020-07-06)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
+## [0.8.1](https://github.com/natura-cosmeticos/natds-js/compare/v0.8.0...v0.8.1) (2020-07-06)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
+# [0.8.0](https://github.com/natura-cosmeticos/natds-js/compare/v0.7.20...v0.8.0) (2020-07-06)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
+## [0.7.20](https://github.com/natura-cosmeticos/natds-js/compare/v0.7.19...v0.7.20) (2020-07-06)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
+## [0.7.19](https://github.com/natura-cosmeticos/natds-js/compare/v0.7.18...v0.7.19) (2020-07-06)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-web
+
+
+
+
+
 ## [0.7.18](https://github.com/natura-cosmeticos/natds-js/compare/v0.7.17...v0.7.18) (2020-07-06)
 
 **Note:** Version bump only for package @naturacosmeticos/natds-web